Move WearableClient into common/
