+++ wear/src/main/java/org/runnerup/service/ListenerService.java
@@ -38,6 +38,7 @@ import org.runnerup.R;
 import org.runnerup.common.tracker.TrackerState;
 import org.runnerup.common.util.Constants;
 import org.runnerup.view.MainActivity;
+import org.runnerup.wear.WearableClient;
 
 public class ListenerService extends WearableListenerService {
